+++ playground/vue-2/src/App.vue
@@ -1,6 +1,8 @@
+
+
 <template>
   <div id="app" lang="en">
-      <link
+    <link
       href="http://localhost:8080/genesys-webcomponents/genesys-webcomponents.css"
       rel="stylesheet"
     />
@@ -10,17 +12,33 @@
       src="http://localhost:8080/genesys-webcomponents/genesys-webcomponents.esm.js"
     ></script>
 
-    <HelloWorld />
+    <!-- <link
+      href="https://dhqbrvplips7x.cloudfront.net/common-ui-docs/genesys-webcomponents/3.41.3-585/genesys-webcomponents/genesys-webcomponents.css"
+      rel="stylesheet"
+    />
+
+    <script
+      type="module"
+      src="https://dhqbrvplips7x.cloudfront.net/common-ui-docs/genesys-webcomponents/3.41.3-585/genesys-webcomponents/genesys-webcomponents.esm.js"
+    ></script> -->
+
+    <Parent />
   </div>
 </template>
 
+<!--
+<template>
+  <div id="app" lang="en">
+    <Parent/>
+  </div>
+</template>
+-->
 <script>
-import HelloWorld from './components/HelloWorld.vue'
-
+import Parent from './components/Parent.vue'
 export default {
   name: 'App',
   components: {
-    HelloWorld
+    Parent
   }
 }
 </script>