+++ src/components/beta/gux-date/tests/gux-date.e2e.ts
@@ -5,10 +5,10 @@ const date = '2022-07-07T13:35:30.100Z';
 describe('gux-date-beta', () => {
   describe('#render', () => {
     [
-        `<gux-date-beta datetime=${date} format="short"></gux-date-beta>`,
-        `<gux-date-beta datetime=${date} format="medium"></gux-date-beta>`,
-        `<gux-date-beta datetime=${date} format="long"></gux-date-beta>`,
-        `<gux-date-beta datetime=${date} format="full"></gux-date-beta>`
+      `<gux-date-beta datetime=${date} format="short"></gux-date-beta>`,
+      `<gux-date-beta datetime=${date} format="medium"></gux-date-beta>`,
+      `<gux-date-beta datetime=${date} format="long"></gux-date-beta>`,
+      `<gux-date-beta datetime=${date} format="full"></gux-date-beta>`
     ].forEach((html, index) => {
       it(`should render component as expected (${index + 1})`, async () => {
         const page = await newSparkE2EPage({ html });