+++ src/components/beta/gux-table/gux-table.tsx
@@ -28,6 +28,10 @@ import {
 
 const COL_RESIZE_HANDLE_WIDTH = 3;
 
+/**
+ * @slot data - Slot for table element
+ */
+
 @Component({
   styleUrl: 'gux-table.less',
   tag: 'gux-table-beta',
@@ -46,7 +50,6 @@ export class GuxTable {
   private tableId: string = randomHTMLId('gux-table');
   private columnsWidths: object = {};
   private tableWidth: number = this.getElementComputedWidth(this.slottedTable);
-
   /**
    * Indicates that vertical scroll is presented for table
    */
@@ -553,8 +556,8 @@ export class GuxTable {
       columnsStyles += `[gs-table-id=${
         this.tableId
       }] th[data-column-name="${column}"]{
-        width:${String(this.columnsWidths[column])};
-        min-width:${String(this.columnsWidths[column])};
+        width:${String(this.columnsWidths[column])}px;
+        min-width:${String(this.columnsWidths[column])}px;
       }`;
     });