+++ src/components/stable/gux-listbox-multi/gux-listbox-multi.tsx
@@ -193,9 +193,9 @@ export class GuxListboxMulti {
     this.listboxOptions.forEach(listboxOption => {
       if (
         listboxOption
-          .querySelector('.gux-option')
-          .textContent.toLowerCase()
-          .trim() == this.textInput.toLowerCase().trim()
+          .querySelector('[gux-slot-container]')
+          ?.textContent?.toLowerCase()
+          ?.trim() == this.textInput.toLowerCase().trim()
       ) {
         hasExactMatch = true;
         this.hasExactMatch = true;
@@ -237,6 +237,7 @@ export class GuxListboxMulti {
     this.listboxOptions = (
       Array.from(this.root.children) as HTMLGuxOptionMultiElement[]
     ).filter(element => element.tagName === 'GUX-OPTION-MULTI');
+    this.internallistboxoptionsupdated.emit();
   }
 
   private updateListboxOptions(): void {
@@ -299,12 +300,14 @@ export class GuxListboxMulti {
   }
 
   renderAllListboxOptionsFiltered(): JSX.Element {
-    return [
-      <div class="gux-message-container">
-        <div class="gux-no-matches">{this.i18n('noMatches')}</div>
-      </div>,
-      this.renderHiddenSlot()
-    ] as JSX.Element;
+    if (this.allListboxOptionsFiltered) {
+      return [
+        <div class="gux-message-container">
+          <div class="gux-no-matches">{this.i18n('noMatches')}</div>
+        </div>,
+        this.renderHiddenSlot()
+      ] as JSX.Element;
+    }
   }
 
   renderCreateOptionSlot(): JSX.Element {
@@ -316,13 +319,10 @@ export class GuxListboxMulti {
       return this.renderLoading();
     }
 
-    if (this.allListboxOptionsFiltered) {
-      return this.renderAllListboxOptionsFiltered();
-    }
-
     return (
       <Host role="listbox" aria-multiselectable="true" tabindex={0}>
         <slot onSlotchange={() => this.updateOnSlotChange()} />
+        {this.renderAllListboxOptionsFiltered()}
         {this.renderCreateOptionSlot()}
       </Host>
     ) as JSX.Element;