+++ src/utils/dom/get-slot.ts
@@ -1,0 +1,6 @@
+export function getSlot(
+  element: HTMLElement,
+  slotName: string
+): HTMLSlotElement {
+  return element.querySelector(`[slot=${slotName}]`);
+}